Move formatDate and formatFullDate functions into separate file
